Refactor: Created a component out of the URL Bar
Moved the child from `FlexBrowserWindow` to its own component `URLBar`
